// src/api/types/CampaignReportAbSplitB.ts
//  This file was auto-generated from our API Definition.

/**
 * Stats for Campaign B.
 */
export interface CampaignReportAbSplitB {
    /** Abuse reports for Campaign B. */
    abuseReports?: number;
    /** Bounces for Campaign B. */
    bounces?: number;
    /** Forwards for Campaign B. */
    forwards?: number;
    /** Opens for forwards from Campaign B. */
    forwardsOpens?: number;
    /** The last open for Campaign B. */
    lastOpen?: string;
    /** Opens for Campaign B. */
    opens?: number;
    /** Recipients clicks for Campaign B. */
    recipientClicks?: number;
    /** Unique opens for Campaign B. */
    uniqueOpens?: number;
    /** Unsubscribes for Campaign B. */
    unsubs?: number;
}